Fill sorts select element with correct data
