+++ frontend/src/Features/ManageEntity/Author/ManageAuthorPage.tsx
@@ -1,12 +1,13 @@
 import {
   Button,
   Checkbox,
-  Divider, Input, InputGroup, InputLeftElement, Select, Table, TableContainer, Tbody, Td, Th, Thead, Tr,
+  Divider, Input, InputGroup, InputLeftElement, Select, Table,
+  TableContainer, Tbody, Td, Th, Thead, Tr,
 } from '@chakra-ui/react';
 import { PencilIcon, SearchIcon, TrashIcon } from '@heroicons/react/outline';
 import { useEffect, useMemo, useState } from 'react';
-import EntityMapper from '../../../Functions/Mappers/EntityMapper';
 import BasePage from '../../Components/BasePage';
+import useManageAuthorViewModel from './ManageAuthorViewModel';
 
 function ManageAuthorHeader() {
   return (
@@ -18,152 +19,8 @@ function ManageAuthorHeader() {
   );
 }
 
-const dummyData = () => ([
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-  {
-    name: 'J.K. Rowling',
-    birth_date: 'July 31, 1965',
-    total_books: 20,
-    total_copies_owned: 82,
-    currently_borrowed: 42,
-  },
-  {
-    name: 'Carl Sagan',
-    birth_date: 'November 9, 1934',
-    total_books: 10,
-    total_copies_owned: 52,
-    currently_borrowed: 50,
-  },
-]);
-
 export default function ManageAuthorPage() {
-  const [authorData, setAuthorData] = useState<any[]>([]);
-
-  useEffect(() => {
-    const dummyAuthors = dummyData().map((element) => ({
-      name: element.name,
-      birth_date: element.birth_date,
-      total_books: element.total_books,
-      total_copies_owned: element.total_copies_owned,
-      currently_borrowed: element.currently_borrowed,
-    }));
-
-    setAuthorData(dummyAuthors);
-  }, []);
-
-  const authorsElement = useMemo(() => authorData.map((element) => (
-    <Tr>
-      <Td>{element.name}</Td>
-      <Td>{element.birth_date}</Td>
-      <Td>{element.total_books}</Td>
-      <Td>{element.total_copies_owned}</Td>
-      <Td>{element.currently_borrowed}</Td>
-      <Td>
-        <div className="flex flex-row items-center gap-6">
-          <PencilIcon className="h-6 w-6 stroke-gray-500" />
-          <TrashIcon className="h-6 w-6 stroke-gray-500" />
-          <Checkbox size="lg" />
-        </div>
-      </Td>
-    </Tr>
-  )), [authorData]);
+  const viewModel = useManageAuthorViewModel();
 
   return (
     <BasePage path="manage-author">
@@ -177,14 +34,15 @@ export default function ManageAuthorPage() {
             <Input type="text" placeholder="J.K. Rowling" />
           </InputGroup>
           <Select placeholder="Sort By" w="20%">
-            <option value="option1">Option 1</option>
-            <option value="option2">Option 2</option>
-            <option value="option3">Option 3</option>
+            <option value="name">Name</option>
+            <option value="birth-date">Birth date</option>
+            <option value="total-books">Total books</option>
+            <option value="total-copies-owned">Total copies owned</option>
+            <option value="currently-borrowed">Currently borrowed</option>
           </Select>
           <Select placeholder="Sort Order" w="20%">
-            <option value="option1">Option 1</option>
-            <option value="option2">Option 2</option>
-            <option value="option3">Option 3</option>
+            <option value="ascending">Ascending</option>
+            <option value="descending">Descending</option>
           </Select>
           <Button colorScheme="blue" width="10%">
             Create
@@ -200,11 +58,11 @@ export default function ManageAuthorPage() {
                 <Th w="13%">Total books</Th>
                 <Th w="13%">Total copies owned</Th>
                 <Th w="13%">Currently borrowed</Th>
-                <Th w="10%" />
+                <Th w="10%">Actions</Th>
               </Tr>
             </Thead>
             <Tbody>
-              {authorsElement}
+              {viewModel.authorsElement}
             </Tbody>
           </Table>
         </TableContainer>
